+++ src/state.rs
@@ -292,6 +292,11 @@ impl<'shell> State<'shell> {
                             "Taylor Cramer <cramertj@google.com>, Benjamin Brittain <bwb@google.com>, Erick Tryzelaar <etryzelaar@google.com>",
                         ),
                         (
+                            "is-terminal",
+                            "MIT",
+                            "softprops <d.tangren@gmail.com>, Dan Gohman <dev@sunfishcode.online>",
+                        ),
+                        (
                             "libsw",
                             "MIT OR Apache-2.0",
                             "Ula Shipman <ula.hello@mailbox.org>",